+++ packages/frontend/vite.config.ts
@@ -8,6 +8,9 @@ module.exports = {
   alias: {
     "/@/": resolve(__dirname, "./src"),
   },
+  define: {
+    wsEndpoint: `ws://${config.server.host}:${config.server.port}/graphql`,
+  },
   proxy: {
     "/graphql": { target },
     "/playground": { target },