+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Graceful "Not found" handler
 - Villus GraphQL client
 - Type-safe GraphQL context
+- Subscriptions client forwarder